% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlockSim-package.R
\docType{package}
\name{BlockSim-package}
\alias{BlockSim-package}
\alias{BlockSim}
\title{BlockSim - Blockchain simulation and research tool}
\description{
BlockSim - Blockchain simulation and research tool
}
\keyword{package}
