% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{complement}
\alias{complement}
\title{Find the complementary set}
\usage{
complement(n, v)
}
\arguments{
\item{n}{The size of complete set}

\item{v}{The vector containing a subset of 1:n}
}
\value{
This function will return the complementary set of a given subset.
}
\description{
Find the complementary set
}
