% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_block_distance}
\alias{estimate_block_distance}
\title{Estimate block distance}
\usage{
estimate_block_distance(block_sequential_distance, block_rate, growth_rate)
}
\description{
Estimate block distance
}
