% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_growth_rate}
\alias{estimate_growth_rate}
\title{Estiamte the growth rate of blockchain}
\usage{
estimate_growth_rate(block_rate, block_size = 1, band_width = 512,
  eps = 1e-100, gamma_shape = 2)
}
\description{
Estiamte the growth rate of blockchain
}
