% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_max_block_discord}
\alias{estimate_max_block_discord}
\title{Estimate maximum block discord}
\usage{
estimate_max_block_discord(block_rate, block_size, band_width = 512,
  gamma_shape = 2, p = 0.9999)
}
\description{
Estimate maximum block discord
}
