% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_observable_probability}
\alias{estimate_observable_probability}
\title{Estimate observable probability}
\usage{
estimate_observable_probability(N, block_rate, block_size,
  band_width = 512, gamma_shape = 2)
}
\description{
Estimate observable probability
}
