% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_reachable_probability}
\alias{estimate_reachable_probability}
\title{Estimate reachable probability}
\usage{
estimate_reachable_probability(
  N,
  block_rate,
  block_size = 1,
  band_width = 512,
  gamma_shape = 2
)
}
\description{
Estimate reachable probability
}
