% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_tip_probability}
\alias{estimate_tip_probability}
\title{Estimate tip probability}
\usage{
estimate_tip_probability(N, block_rate, block_size, band_width = 512,
  gamma_shape = 2)
}
\description{
Estimate tip probability
}
