% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{shortest_distances}
\alias{shortest_distances}
\title{Calculate shortest distances of unweighted graph}
\usage{
shortest_distances(adj_matrix, source_nodes = rep_len(TRUE,
  dim(adj_matrix)[1]))
}
\arguments{
\item{adj_matrix}{Logical adjacency matrix of given unweighted graph}

\item{source_nodes}{Logical vector to indicate the source nodes that 
need to calculate the shortest distances}
}
\value{
This function will return the shortest distance matrix, where the element
\code{[i, j]} is the shortest distance between node i and j. Value Inf means unreachable.
If \code{source_nodes[i]} equals FALSE, the shortest distance from i to other nodes
will not be calculated and the row i will be all Inf.
}
\description{
Calculate all pairs of shortest distances of unweighted graph
}
\details{
This function calculates all pairs of shortest distances of unweighted graph
by using breadth-first-search (BFS) algorithm.
}
