% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_blockchain}
\alias{simulate_blockchain}
\title{Simulate a blockchain}
\usage{
simulate_blockchain(
  n,
  attack_power = 0,
  t1 = ceiling(1/3 * n),
  t2 = ceiling(2/3 * n),
  n_miners = 10000,
  block_rate = 1/600,
  block_size = 1,
  n_genesis_blocks = 1,
  burn_in = n_genesis_blocks,
  fixed_block_interval = TRUE
)
}
\arguments{
\item{n}{the size of blockchain, i.e. the total number of blocks in the blockchain}

\item{attack_power}{the percentage of attacker's computing power}

\item{t1}{the first time point of attack}

\item{t2}{the second time point of attack}

\item{n_miners}{the number of miners}

\item{block_rate}{the block genration rate (blocks per sec)}

\item{block_size}{the block size (MB)}

\item{burn_in}{the burn in time points, in which period all blocks are generated by the genesis miner}
}
\value{
This function will return a list with the following components:
  \item{adj_mat}{The adjacent matrix of generated blockchain.}
  \item{block_info}{A list that contains information of blockchain and blocks.}
}
\description{
Simulate a blockchain
}
