% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_block_distance}
\alias{estimate_block_distance}
\title{Estimate block distance}
\usage{
estimate_block_distance(
  N,
  block_rate,
  block_size = 1,
  band_width = 512,
  gamma_shape = 2
)
}
\description{
Estimate block distance
}
